feat(read-and-create-calendar-events): merge node examples
